+++ src/architecture/msgPayloadDefC/THRConfigMsgPayload.h
@@ -24,9 +24,9 @@
 
 /*! @brief Single Thruster configuration message */
 typedef struct {
-    double rThrust_B[3];        //!< [m] Location of the thruster in the spacecraft
-    double tHatThrust_B[3];     //!< [-] Unit vector of the thrust direction
-    double maxThrust;			//!< [N] Max thrust
+    float rThrust_B[3];        //!< [m] Location of the thruster in the spacecraft
+    float tHatThrust_B[3];     //!< [-] Unit vector of the thrust direction
+    float maxThrust;			//!< [N] Max thrust
 }THRConfigMsgPayload;
