+++ src/fswAlgorithms/effectorInterfaces/thrForceMapping/thrForceMapping.h
@@ -33,19 +33,19 @@
 /*!@brief Data structure for module to map a command torque onto thruster forces. */
 typedef struct {
     /* declare module public variables */
-    double   controlAxes_B[3*3];                    //!< []      array of the control unit axes
-    double   rThruster_B[MAX_EFF_CNT][3];           //!< [m]     local copy of the thruster locations
-    double   gtThruster_B[MAX_EFF_CNT][3];          //!< []      local copy of the thruster force unit direction vectors
+    float   controlAxes_B[3*3];                    //!< []      array of the control unit axes
+    float   rThruster_B[MAX_EFF_CNT][3];           //!< [m]     local copy of the thruster locations
+    float   gtThruster_B[MAX_EFF_CNT][3];          //!< []      local copy of the thruster force unit direction vectors
     int32_t  thrForceSign;                          //!< []      Flag indicating if pos (+1) or negative (-1) thruster solutions are found
-    double angErrThresh;                            //!< [r]     Angular error at which thruster forces are scaled to not be super-saturated
-    double   epsilon;                               //!<         variable specifying what is considered a small number
+    float angErrThresh;                            //!< [r]     Angular error at which thruster forces are scaled to not be super-saturated
+    float   epsilon;                               //!<         variable specifying what is considered a small number
     uint32_t use2ndLoop;                            //!< []      flag indicating if the 2nd least squares fitting loop should be used (1) or not used (0 - default)
 
     /* declare module private variables */
     uint32_t numControlAxes;                        //!< []      counter indicating how many orthogonal axes are controlled
     uint32_t numThrusters;                          //!< []      The number of thrusters available on vehicle
-    double outTorqAngErr;                           //!< [r]     Angular error of effector torque
-    double thrForcMag[MAX_EFF_CNT];                 //!<         vector of thruster force magnitudes
+    float outTorqAngErr;                           //!< [r]     Angular error of effector torque
+    float thrForcMag[MAX_EFF_CNT];                 //!<         vector of thruster force magnitudes
 
     /* declare module IO interfaces */
     THRArrayCmdForceMsg_C thrForceCmdOutMsg;        //!< The name of the output thruster force message
@@ -66,11 +66,11 @@ extern "C" {
     void Update_thrForceMapping(thrForceMappingConfig *configData, uint64_t callTime, int64_t moduleID);
     void Reset_thrForceMapping(thrForceMappingConfig *configData, uint64_t callTime, int64_t moduleID);
 
-    void substractMin(double *F, uint32_t size);
+    void substractMin(float *F, uint32_t size);
     void findMinimumNormForce(thrForceMappingConfig *configData,
-                              double D[3][MAX_EFF_CNT], double Lr_B[3], uint32_t numForces, double F[MAX_EFF_CNT]);
-    double computeTorqueAngErr(double D[3][MAX_EFF_CNT], double BLr[3], uint32_t numForces, double epsilon,
-                               double F[MAX_EFF_CNT], double FMag[MAX_EFF_CNT]);
+                              float D[3][MAX_EFF_CNT], float Lr_B[3], uint32_t numForces, float F[MAX_EFF_CNT]);
+    float computeTorqueAngErr(float D[3][MAX_EFF_CNT], float BLr[3], uint32_t numForces, float epsilon,
+                               float F[MAX_EFF_CNT], float FMag[MAX_EFF_CNT]);
 
 #ifdef __cplusplus
 }
